Fixa lite varningar

+++ src/skyflower/Components/Event.cpp
@@ -356,7 +356,7 @@ int Event::SetSpeed(lua_State* L)
 
 		Entity* entityAi = entityManager->getEntity(aiId);
 
-		entityAi->getComponent<Movement*>("Movement")->SetSpeed(lua_tonumber(L, 2));
+		entityAi->getComponent<Movement*>("Movement")->SetSpeed((float)lua_tonumber(L, 2));
 	}
 
 	return 0;
